CostPersonal.java
 
 /**
* @author Hugo Martin 
* @version 1.0
* @since 30/01/2023
*/public class CostPersonal {
        // static enum TipusTreballador{
        // DIRECTOR, SUBDIRECTOR, BASE
        // }
        static float CalculaCostDelPersonal(Treballador treballadors[]) {
        float costFinal = 0;
        Treballador treballador;
        for (int i = 0; i < treballadors.length; i++) {
        treballador = treballadors[i];
        if (treballador.getTipusTreballador() == Treballador.DIRECTOR ||
treballador.getTipusTreballador() == Treballador.SUBDIRECTOR) {
    costFinal += treballador.getNomina();
} else {
costFinal += treballador.getNomina() + (treballador.getHoresExtres() * 20);
}
}
return costFinal;
}
static float CostTreballador(Treballador treballador) {
if (treballador.getTipusTreballador() == Treballador.DIRECTOR || treballador.getTipusTreballador() == Treballador.SUBDIRECTOR) {
return treballador.getNomina();
} else {
    return treballador.getNomina() +
    (treballador.getHoresExtres() *
    20);
    }
    }
    static float CalculaCostDelPersonal2(Treballador treballadors[]) {
    // Versio refactoritzada del mètode anterior
    float costFinal = 0;
    Treballador treballador;
    for (int i = 0; i < treballadors.length; i++) {
    costFinal = costFinal + CostTreballador(treballadors[i]);
    }
    return costFinal;
    }
    }

